STORAGE: Rename package to holostorageaccessor
