+++ HoloStorageAccessor/internal/holostorageaccessor/api_default.go
@@ -6,7 +6,7 @@
  * API version: 1.0.0
  */
 
-package openapi
+package holostorageaccessor
 
 import (
 	"bytes"
