+++ HoloStorageAccessor/internal/holostorageaccessor/blob_storage_client.go
@@ -1,4 +1,4 @@
-package openapi
+package holostorageaccessor
 
 import (
 	"bytes"
