+++ HoloStorageAccessor/internal/holostorageaccessor/model_author.go
@@ -6,7 +6,7 @@
  * API version: 1.0.0
  */
 
-package openapi
+package holostorageaccessor
 
 // Author - Who authored the Hologram
 type Author struct {
