+++ HoloStorageAccessor/internal/holostorageaccessor/model_author_test.go
@@ -1,4 +1,4 @@
-package openapi
+package holostorageaccessor
 
 import (
 	"testing"
