+++ HoloStorageAccessor/internal/holostorageaccessor/model_hologram.go
@@ -6,7 +6,7 @@
  * API version: 1.0.0
  */
 
-package openapi
+package holostorageaccessor
 
 import (
 	"encoding/json"
