+++ HoloStorageAccessor/internal/holostorageaccessor/model_hologram_upload_data_test.go
@@ -6,7 +6,7 @@
  * API version: 1.0.0
  */
 
-package openapi
+package holostorageaccessor
 
 import (
 	"testing"
