+++ HoloStorageAccessor/internal/holostorageaccessor/utils.go
@@ -1,4 +1,4 @@
-package openapi
+package holostorageaccessor
 
 import (
 	"encoding/json"
